requirements.txt
fastapi
uvicorn
faiss-cpu
python-dotenv
sentence-transformers
openai
